handle logout header

+++ FrontEnd/static/js/api/logout.js
@@ -1,5 +1,19 @@
-document.getElementById("logout").addEventListener("click", (e)=> {
-    fetch("api/logout", {
-        method:"DELETE",
+function DeleteHeader() {
+    if (document.cookie.includes("token=")) {
+        const nav = document.getElementById("nav")
+        nav.innerHTML = '<p>Welcome To Forum |</p> <a id="logout">Logout</a>'
+    }
+}
+
+DeleteHeader()
+const LogoutButton = document.getElementById("logout")
+
+if (LogoutButton) {
+    LogoutButton.addEventListener("click", (e) => {
+        fetch("api/logout", {
+            method: "DELETE",
+        })
+        window.location.href = "/"
     })
-})
+}
+
